[Admin][Taxon][Behat] Separate TreeElement from CreatePage
