+++ src/Sylius/Behat/Resources/config/services/contexts/ui.xml
@@ -305,6 +305,7 @@
             <argument type="service" id="sylius.behat.page.admin.taxon.update" />
             <argument type="service" id="sylius.behat.element.admin.taxon.form" />
             <argument type="service" id="sylius.behat.element.admin.taxon.image_form" />
+            <argument type="service" id="sylius.behat.element.admin.taxon.tree" />
             <argument type="service" id="sylius.behat.notification_checker.admin" />
             <argument type="service" id="sylius.behat.java_script_test_helper" />
             <argument type="service" id="sylius.behat.page.admin.product.update_simple" />
