+++ src/Sylius/Behat/Resources/config/services/elements/admin.xml
@@ -99,6 +99,12 @@
         />
 
         <service
+            id="sylius.behat.element.admin.taxon.tree"
+            class="Sylius\Behat\Element\Admin\Taxon\TreeElement"
+            parent="sylius.behat.element"
+        />
+
+        <service
             id="sylius.behat.element.admin.promotion_coupon.form"
             class="Sylius\Behat\Element\Admin\PromotionCoupon\FormElement"
             parent="sylius.behat.element.admin.crud.form"
